requirements.txt
torch>=2.3.0
transformers>=4.46.0
peft>=0.12.0
accelerate>=0.34.0
bitsandbytes>=0.43.0
datasets>=3.0.0
huggingface_hub>=0.25.0
anthropic>=0.36.0
google-genai>=1.0.0
openai>=1.52.0
pillow>=10.0.0
matplotlib>=3.9.0
pandas>=2.2.0
numpy>=1.26.0
tqdm>=4.66.0
python-dotenv>=1.0.0
jsonlines>=4.0.0
opencv-python>=4.10.0
einops>=0.8.0
qwen-vl-utils>=0.0.8
# Depth Anything V2 (optional Phase 5 ablation) installed from source — see README
